Consolidated testing code in test/support; removed elixir_ale dependency
